perf(module/config): seperate config
